Implement feat to update user password

+++ backend/sigma/src/main/java/com/klikk/sigma/entity/Checkout.java
@@ -46,7 +46,7 @@ public class Checkout {
     private String billingState;
 
     @Column(name = "postcode")
-    private Integer postcode;
+    private Long postcode;
 
     @Column(name = "country")
     private String billingCountry;
